% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_origin_of_species.R
\name{TheOriginOfSpecies}
\alias{TheOriginOfSpecies}
\title{Create a phylogeny}
\usage{
TheOriginOfSpecies(size, start)
}
\arguments{
\item{size}{The size of the phylogeny (number of tips).}

\item{start}{The tip label (number) used for the start.}
}
\description{
Create a phylogeny
}
